file updated

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,6 @@
 # Executes a bash command
 exec { 'kill':
   command => 'pkill -f killmenow',
-  path    => ['/usr/bin', '/usr/sbin']
+  path    => ['/usr/bin', '/usr/sbin'],
+  onlyif  => 'pgrep -f killmenow',
 }
